Fix -Wall -Wextra -Werror warnings
- envelope.c: include face_tss/config.h for face_tss_normalize_name
  declaration; cast uint16_t to uint32_t in sign-compare checks
- transport.c: include flatcc/flatcc_builder.h for flatcc_builder_free
- tss.c: replace strncpy with memcpy + explicit NUL terminator
- types.h / types.hpp: guard FACE_INF_TIME_VALUE with #ifndef (defined
  in both C and C++ headers with identical values)

+++ c/src/envelope.c
@@ -9,6 +9,7 @@
  */
 
 #include "face_tss/envelope.h"
+#include "face_tss/config.h"
 
 #include <stdlib.h>
 #include <string.h>
@@ -349,7 +350,7 @@ FACE_TSS_RETURN_CODE face_tss_envelope_decode(
         uint16_t _e = vt_entry(buf, vtable, vsize, (fid));               \
         if (_e != 0) {                                                   \
             uint32_t _at;                                                \
-            if (_e + 8 > tsize)                                          \
+            if ((uint32_t)_e + 8 > tsize)                                \
                 return FACE_TSS_RC_INVALID_PARAM;                        \
             _at = table + _e;                                            \
             if (_at > (uint32_t)(len - 8))                               \
@@ -376,7 +377,7 @@ FACE_TSS_RETURN_CODE face_tss_envelope_decode(
     {
         uint16_t _e = vt_entry(buf, vtable, vsize, F_PAYLOAD);
         if (_e != 0) {
-            if (_e + 4 > tsize)
+            if ((uint32_t)_e + 4 > tsize)
                 return FACE_TSS_RC_INVALID_PARAM;
             foff = table + _e;
         }
